refactor: rename user and auth services
