+++ src/app/api/auth/login/route.ts
@@ -4,9 +4,9 @@ import { NextResponse } from 'next/server'
 import { z, ZodError } from 'zod'
 import { NotFound } from 'payload'
 
-import AuthDataService from '@/data-layer/services/AuthService'
+import AuthDataService from '@/data-layer/services/AuthDataService'
 import AuthService from '@/business-layer/services/AuthService'
-import UserService from '@/data-layer/services/UserService'
+import UserDataService from '@/data-layer/services/UserDataService'
 import { cookies } from 'next/headers'
 import { AUTH_COOKIE_NAME } from '@/types/Auth'
 
@@ -17,14 +17,14 @@ export const LoginRequestBodySchema = z.object({
 export type LoginRequestBody = z.infer<typeof LoginRequestBodySchema>
 
 export const POST = async (req: NextRequest) => {
-  const authDataService = new AuthDataService()
   const authService = new AuthService()
-  const userService = new UserService()
+  const authDataService = new AuthDataService()
+  const userDataService = new UserDataService()
 
   try {
     const { email, password } = LoginRequestBodySchema.parse(await req.json())
     const auth = await authDataService.getAuthByEmail(email)
-    const user = await userService.getUserByEmail(email)
+    const user = await userDataService.getUserByEmail(email)
 
     const isValid = await authService.verifyPassword(password, auth.password || '')
     if (!isValid) {
